# backend/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Clarift API",
    "version": "1.0.0"
  },
  "paths": {
...
    "/api/v1/quizzes": {
      "get": {
        "tags": [
          "quizzes"
        ],
        "summary": "List Quizzes",
        "operationId": "list_quizzes_api_v1_quizzes_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/QuizItemResponse"
                  },
                  "type": "array",
                  "title": "Response List Quizzes Api V1 Quizzes Get"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      },
      "post": {
        "tags": [
          "quizzes"
        ],
        "summary": "Create Quiz",
        "description": "Request a quiz for a given document.\n\nCreates a pending Quiz record, then triggers the quiz chain.\nReturns the quiz ID and weak topics.",
        "operationId": "create_quiz_api_v1_quizzes_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateQuizRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CreateQuizResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/api/v1/quizzes/{quiz_id}": {
      "get": {
        "tags": [
          "quizzes"
        ],
        "summary": "Get Quiz",
        "operationId": "get_quiz_api_v1_quizzes__quiz_id__get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "quiz_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Quiz Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/QuizDetailResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/quizzes/attempts": {
      "post": {
        "tags": [
          "quizzes"
        ],
        "summary": "Submit Attempt",
        "description": "Submit a quiz attempt and get results.\n\nCreates a QuizAttempt record, calculates score, updates topic performance.\nReturns score and weak topics.",
        "operationId": "submit_attempt_api_v1_quizzes_attempts_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SubmitAttemptRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SubmitAttemptResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "CreateQuizRequest": {
        "properties": {
          "document_id": {
            "type": "string",
            "title": "Document Id"
          },
          "question_count": {
            "type": "integer",
            "title": "Question Count",
            "default": 5
          },
          "auto_mode": {
            "type": "boolean",
            "title": "Auto Mode",
            "default": true
          }
        },
        "type": "object",
        "required": [
          "document_id"
        ],
        "title": "CreateQuizRequest",
        "description": "Request body for creating a quiz."
      },
      "CreateQuizResponse": {
        "properties": {
          "quiz_id": {
            "type": "string",
            "title": "Quiz Id"
          },
          "message": {
            "type": "string",
            "title": "Message"
          },
          "weak_topics": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Weak Topics"
          }
        },
        "type": "object",
        "required": [
          "quiz_id",
          "message",
          "weak_topics"
        ],
        "title": "CreateQuizResponse",
        "description": "Response after creating a quiz."
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "QuizDetailResponse": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "document_id": {
            "type": "string",
            "title": "Document Id"
          },
          "questions": {
            "items": {
              "additionalProperties": true,
              "type": "object"
            },
            "type": "array",
            "title": "Questions"
          },
          "question_types": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Question Types"
          },
          "question_count": {
            "type": "integer",
            "title": "Question Count"
          },
          "auto_mode": {
            "type": "boolean",
            "title": "Auto Mode"
          },
          "created_at": {
            "type": "string",
            "title": "Created At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "document_id",
          "questions",
          "question_types",
          "question_count",
          "auto_mode",
          "created_at"
        ],
        "title": "QuizDetailResponse"
      },
      "QuizItemResponse": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "document_id": {
            "type": "string",
            "title": "Document Id"
          },
          "question_count": {
            "type": "integer",
            "title": "Question Count"
          },
          "question_types": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Question Types"
          },
          "created_at": {
            "type": "string",
            "title": "Created At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "document_id",
          "question_count",
          "question_types",
          "created_at"
        ],
        "title": "QuizItemResponse"
      },
      "SubmitAttemptRequest": {
        "properties": {
          "quiz_id": {
            "type": "string",
            "title": "Quiz Id"
          },
          "answers": {
            "additionalProperties": {
              "type": "string"
            },
            "type": "object",
            "title": "Answers"
          }
        },
        "type": "object",
        "required": [
          "quiz_id",
          "answers"
        ],
        "title": "SubmitAttemptRequest",
        "description": "Request body for submitting a quiz attempt."
      },
      "SubmitAttemptResponse": {
        "properties": {
          "attempt_id": {
            "type": "string",
            "title": "Attempt Id"
          },
          "score": {
            "type": "number",
            "title": "Score"
          },
          "weak_topics": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Weak Topics"
          },
          "message": {
            "type": "string",
            "title": "Message"
          }
        },
        "type": "object",
        "required": [
          "attempt_id",
          "score",
          "weak_topics",
          "message"
        ],
        "title": "SubmitAttemptResponse",
        "description": "Response after submitting an attempt."
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    },
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  }
}